+++ src/ArrayFunctions/AFGet.php
@@ -17,14 +17,32 @@ class AFGet extends ArrayFunction {
 	 * @inheritDoc
 	 */
 	public function execute( array $value, string ...$indices ): array {
-		foreach ( $indices as $index ) {
-			if ( !is_array( $value ) || !isset( $value[$index] ) ) {
-				return [ '' ];
-			}
+		return [ $this->index( $value, $indices ) ?? '' ];
+	}
 
-			$value = $value[$index];
+	/**
+	 * Index the given array using the list of indices, with support for wildcards.
+	 *
+	 * @param mixed $value
+	 * @param array $indices
+	 * @return mixed
+	 */
+	private function index( $value, array $indices ) {
+		if ( empty( $indices ) ) {
+			return $value;
 		}
 
-		return [ $value ];
+		if ( !is_array( $value ) ) {
+			// Not indexable
+			return null;
+		}
+
+		$index = array_shift( $indices );
+
+		if ( isset( $value[$index] ) ) {
+			return $this->index( $value[$index], $indices );
+		} else {
+			return null;
+		}
 	}
 }